Support uniquely named anchors even with duplicate text
Summary:
The first incarnation of the anchors didn't allow for unique anchors for
equivalent text.

This does.

If we design our docs right, there should only be one h1 always. So no need
to worry there. Text for H2 should be unique too, otherwise, what's going on?
It is when we get to H3, that we can have duplicate text. So this handles that.

I don't like more than three levels of headings in any doc. So I didn't go
further than the third level here. If we have an H4, then it won't be
uniquely anchored with this, but shouldn't be too hard to add.

Reviewed By: ssorallen

Differential Revision: D2865567

fb-gh-sync-id: ce1aeed6e2a6125bb65d6e86ef221da94520db08