Remove automatic kramdown id suffixes
Summary:
Kramdown automatically suffixes ids with `-1`, `-2`, etc. when it
encounters a heading name that has existed before.

We don't want that since we are using a tree-style format for our
anchoring, a la `h2Name__h3Name`, etc.

So as we have it now, current links can be invalid with the move
to Kramdown. This fixes that problem by stripping out the auto
suffix from the id attribute.

Reviewed By: ssorallen

Differential Revision: D2897363

fb-gh-sync-id: e5bc532dbfb96bbe73d42b635d87fc39838d7b89