Keybindings (7/N): Common Keyboard Shortcuts
Summary:
I like to have a table of quick access, common shortcuts near the top. This takes a few of the shortcuts in the various sections below and brings them to this table.

The words there were here I am going to utilize in some fashion in an `Editor Basics` section which is coming up soon.

Facebook

Depends on D2908297

Reviewed By: jamesgpearce

Differential Revision: D2908476

fb-gh-sync-id: 943ec730466a5d6ff10c51431d2d7aa46e328c6a
shipit-source-id: 943ec730466a5d6ff10c51431d2d7aa46e328c6a