Add table of contents to most pages
Summary:I have wanted this for a while. Found out Kramdown has a feature
to easily add a table of contents based on header ids (which we
have). So I add the two-line code block to most of our markdown
pages. (Not the lanugages one yet, since I am going to make some
modifications to those, and not the uninstall since it didn't
seem right for something that short).

Reviewed By: jamesgpearce

Differential Revision: D2971956

fb-gh-sync-id: abe9e2d86e56ebc35af1484fc50b12135eaa142d
shipit-source-id: abe9e2d86e56ebc35af1484fc50b12135eaa142d