Stop using animated gifs
Summary:Having plain images are easier to deal with. Both maintainence wise
and look and feel wise.

So I am removing the animated gifs from nuclide.io in favor of
retina images.

I think they at least convey good information given that the animated
gifs were hard to see anyway.

And we will be able to save some MB as well in the end when optimize
the entire lot.

Reviewed By: jamesgpearce

Differential Revision: D3034206

fb-gh-sync-id: a295c27883308e97ffdc53f1bba7f4ae486433d2
shipit-source-id: a295c27883308e97ffdc53f1bba7f4ae486433d2