Restructure Debugging (1/N): Hack and PHP
Summary:I believe the platform and language debugging information should be
in their respective sections. The Debugger feature section will have
the basics (using Node debugging as an example). Then link to the
platform and language specific debugging sections elsewhere.

This is the first of a stack of these type of diffs.

Reviewed By: bottledwalter, johnislarry

Differential Revision: D3041300

fb-gh-sync-id: 62c97afd5de15bf1fdebe1d0e31d5f9bf8fef5b1
shipit-source-id: 62c97afd5de15bf1fdebe1d0e31d5f9bf8fef5b1