Fix broken promo links
Summary:The front page redesign that landed today did not have working promo links. Go to nuclide.io and try
clicking on "Get Started" for example.

1. There were no links in the `a href` specified.
2. They were not clickable regardless. I think they were being
hidden by a parent div or something. Not sure. I mucked with the CSS
a bit to get it to work. Let me know if that was the right fix.

Reviewed By: alsutton

Differential Revision: D3063710

fb-gh-sync-id: 594b20732bb92b2931b9d27ce3eb0096db8ae81e
shipit-source-id: 594b20732bb92b2931b9d27ce3eb0096db8ae81e