Show loading spinner while waiting for thread info in the debugger
Summary: When switching from Running to Paused, we expect thread info from the debugger backends. For apps with many threads, this can take a while to produce - and currently the threads pane looks like it's finished rendering, but it shows no threads.  This adds a loading spinner to indicate thread info isn't ready yet.  This helps a lot in the Java case for a large app like fb4a that has many threads.

Reviewed By: johnislarry

Differential Revision: D4615638

fbshipit-source-id: ec76b25af463e1fb3e9d950e53936bb7aba4f0a2