Debugger UX polish: Scroll selected thread into view when breaking
Summary: Currently when the debugger hits a breakpoint or async-breaks, if the attached app has more threads than fit into the thread window's view port, the selected thread can be out of the visible region and hard to find. This change scrolls it into view if needed whenever the debugger stops.

Reviewed By: johnislarry

Differential Revision: D4623269

fbshipit-source-id: 1ff2e90430c62f088388248132e32cde2fd9ffc7