Prettifying diff/patch view
Summary:
Added a color scheme for comments.  Making .diff files use this color scheme (all of them, for now)

I want to make this so that only the '.diff' files opened from certain mercurial commands will use this logic, as we want to be able to draw over these specific views as well, not just any file that happens to have a .diff filetype.  Not quite sure how to do this yet

Reviewed By: jgebhardt

Differential Revision: D4625141

fbshipit-source-id: 1432e9dbdcf22d5dcc90c8f7b4979f80cade72ff