Fix unresolvable breakpoints always show as resolved if set after debugger is attached
Summary:
The current communication between the hidden Chrome dev tools and the Nuclide debugger frontend doesn't communicate if a newly added breakpoint is actually resolved or not. As a result, breakpoints added after the debugger is launched/attached are always showing as resolved in the UI, even when they are not resolved. This is bad because the user expects the breakpoint to hit, and it never will.

This change propagates an additional BreakpointAdded event from the dev tools after the debugger resolves the breakpoint, and has the Nuclide UI assume all breakpoints are unresolved until the backend says otherwise.

NOTE: There is/was a hack (no pun intended) for the PHP/Hack debugger that causes all breakpoints to appear resolved when debugging the web server. This remains necessary because the server actually has no idea which source files are in scope since they are loaded lazily.  This change does not modify the behavior for PHP/Hack, but fixes it for the other target types.

Reviewed By: jgebhardt

Differential Revision: D4663576

fbshipit-source-id: 19b0518f44fd4eeb9d759b7ac727457eb8e3b568