LSP implementation of TypeHint
Summary:
Hooks up the Nuclide "TypeHint" feature to the LSP "textDocument/Hover" action.

There are quite a few TODOs left -- LSP envisages a more fully-featured typehint than VSCode supports!

(I haven't yet tested whether this "range" stuff works. The LSP idea is that when you hover over a symbol, it returns the range of that symbol to highlight. At the moment the Hack service doesn't return a range. So I'd have to change Hack to check whether it works.)

Reviewed By: nmote

Differential Revision: D4689193

fbshipit-source-id: 184dab66f1aa5d3fd641461858582008bc56fce4