Remove "useIdeConnection" from nuclide-hack codebase
Summary:
The code used to use "IDE connection" if either (1) you had enabled the setting, or (2) you passed a gatekeeper check. Now the feature's sufficiently mature and tested that it's time to retire those tests, and make everyone use the IDE connection.

(Note: I want to add a new gatekeeper+config setting called "UseLspConnection", but that will come in the future, and I wanted to remove this one first. It would look odd to add that one without first removing this one.)

Reviewed By: nmote

Differential Revision: D4689531

fbshipit-source-id: 57c12af6c7f5b37d2ad7f39811a1a0fe2fed79b8