Beginning to put together the redux structure for interactive commit
Summary:
Added redux modules with a couple basic actions
Moved checkbox generation to a factory method which is passed in to the FileChanges component.  This allows us to customize a lot of functionality of the checkboxes without bloating the FileChanges component.  This will be expanded upon in the next diffs where I start working out the data model better

Depends on D4676354

Reviewed By: aadisriram

Differential Revision: D4678025

fbshipit-source-id: 8059ff3c5731b939e429b306c44feb41e0f67c65