Prefer IPv6 addresses throughout the remote connection stack
Summary: When resolving the IP address during the initial SSH connection, save the IPv4/6 status and thread that through to the `ServerConnection` to ensure that we use the same one consistently. Since the SSH connection has code to manually prefer IPv6 addresses, this basically ensures that we prefer IPv6 addresses through our entire remote connection stack.

Reviewed By: shushz

Differential Revision: D4706959

fbshipit-source-id: dac7b1e67dbcfa528873f07abb0825e144bc3c3e