Checkbox action reducer
Summary:
Update the state when a check action is carried out

Right now, this is pretty inefficient due to (I think) trying to separate the checkbox creation from the FileChanges component, and react having a tough time trying to figure out when (not) to re-render.  I will team up with jgebhardt to figure out a better way to structure this for higher performance.

Depends on D4707072

Reviewed By: aadisriram

Differential Revision: D4710438

fbshipit-source-id: 1a7a96fd142e05e21b8ad69b46d9a34a7eb6078e