Refactor ValueComponent to handle state transitions gracefully
Summary:
The existing state transitions on `ValueComponent` assume that the same backing data is tied to an instance of a component. However with windowing, the `RecordView` and its sub-components (which includes `ValueComponent`) are ephemeral because the backing data is cycled through the already rendered components as you scroll.

All that said, the `ValueComponent` has state associated with it that was not being updated correctly when the `props` changed. These changes refactor how the state of the component is computed and then implements `componentDidUpdate` to ensure that the state is updated correctly when the props change.

Reviewed By: matthewwithanm

Differential Revision: D4654165

fbshipit-source-id: cbbd36f007c933264325277ead445ac3b450c9a5