Report Flow server status in busy signal
Summary:
With push diagnostics, we can no longer use the heuristic of "is there an outstanding diagnostics request" to decide whether to spin the busy signal, since the concept of a request for diagnostics no longer exists. Instead we have to base it on what we know about the Flow server's state.

In the long run this is the right move. Unfortunately right now we have imperfect information. We only learn that the Flow server is busy when a request is sent -- and even then there is roughly a 1 second delay. This should be mostly fine with non-push diagnostics since we send a request for diagnostics (so we will quickly learn when the server is busy). However, with push diagnostics, there is no such request so we have to rely on other requests. If the user is typing or mousing around, we are sending requests (autocomplete and type-at-pos) anyway, but if they aren't it is possible we will never learn that the server is busy.

To rectify this, I will change Flow so that it notifies us via the persistent connection when it goes off on a recheck. That way we can consistently and immediately update the server status so that the busy signal appears correctly.

Reviewed By: zertosh

Differential Revision: D4743971

fbshipit-source-id: add593cab40a5c561057c58fe48ab174e1ad2732