Large file transfer
Summary: Add two APIs that can be chained together for splitting up a file to 65kB messages and sent across RPC calls, which don't completely block the runloop. I have this working in a feature I am working on and there are no visible performance issues when transferring the file.

Reviewed By: hansonw

Differential Revision: D4701201

fbshipit-source-id: 5e4f4546444111d2903c8313338dde95265e4a8b