Ugly hack to copy & paste url
Summary:
Unfortunately, AtomInput when disabled preventsDefault the copy event. I digged a bit but haven't figured out how to prevent that.

A hacky fix is to make the component not disabled but a controlled input and whenever it changes, then re-update the value to the one we want. Unfortunately, this isn't working. I'm not really sure why.

But, what is working is to manually set the value onDidChange (should be called onChange!). Just need to not trigger it if the value is the same otherwise it stack overflows.

I'm not super happy about this but it solves a big problem people have so I think it's worth putting this small hack in.

Reviewed By: azich

Differential Revision: D4763284

fbshipit-source-id: 79b6cd0ee3a08a6c5b2c5f999fb9ecf663a7c9e1