Don't send updates when the CWD stays the same
Summary: The atom path change callback triggers if you remove a project, but the CWD doesn't change in that case. No need to push a value to the observable in that case. This seems to have been intended in the first place, but the Directory object is a new instance, so the `distinctUntilChanged` call would always be ignored.

Reviewed By: hansonw

Differential Revision: D4729646

fbshipit-source-id: ef51bc92ac2d568e323e9e2e115ee41eebbea31e