Make HackSymbolSearch tab only show up for hack projects
Summary:
In previously making HackSymbolSearch a global quick-open provider, that change had caused its tab to always appear in the Quick-Open dialog.

With this current diff we restore the old behavior, where the HackSymbolSearch only appears if there are indeed Hack projects open. (In particular, amongst all the HackServices that nuclide client has a connection to, at least one of them must have a connection to a HackProcess).

Reviewed By: nmote

Differential Revision: D4720726

fbshipit-source-id: 88b87d0a1137ef9627d21b95c6645315db7614ea