AutocompleteCacher: make the nullability of getSuggestions part of the signature
Summary:
When using the autocomplete cacher, part of the contract is that you can return a `null` to avoid caching an empty array.
However, we shouldn't have to make the entire cacher type `T` an `?Array<Type>` since `updateResults` never passes a nullable.

This is a more accurate representation of what we want.

Reviewed By: nmote

Differential Revision: D4760257

fbshipit-source-id: d044c7896f06f758f298f6285dc4f966fec9473a