Loading spinner in threads pane for C++ debugger spins forever
Summary: The loading spinner displays in the UX for the threads window after recieving the Debugger.Paused notification from the chrome protocol and dismisses when it receives the Threads Updated notification. The C++ debugger python backend is sending these events in the opposite order (updating the threads info, then sending Debugger.Paused). With that ordering on the wire, the frontend has no way of knowing when it's OK to dismiss the spinner...

Reviewed By: johnislarry

Differential Revision: D4772358

fbshipit-source-id: ee9d7e8824404aebbca5e14234ac4d32415c3d94