hide tooltips onscroll
Summary:
There's currently a bug where tooltips get stuck when you scroll while a tooltip is visible. The only way to get them unstuck is to find the exact line that triggered them and mouse over it and out again.

This bug is bad enough that I believe it justifies the hacks...

Reviewed By: matthewwithanm

Differential Revision: D4363959

fbshipit-source-id: 08f1165106c8d0ea42635a52e5c61329bf589d57