Stop trying to start the IDE connection after disposal
Summary:
Trying to restart Atom when Flow was initializing would cause Atom to freeze.
The reason for this is that the Flow server state `Subject` gets completed upon
disposal. This means that `FlowProcess::_serverIsReady()` (which I was relying
on to throttle startup attempts) returns immediately, and we end up in an
infinite loop here.

This moves up the disposal check so that it occurs after every asynchronous call
in the loop. This solves the immediate issue and it's what should be done
anyway. Proper throttling (both for attempting to establish a connection and
restarting it when it dies) is still on my to do list.

Reviewed By: hansonw

Differential Revision: D4774759

fbshipit-source-id: 13bf7fa1ac7437ae05070f9f061dd1abee5d2ec1