Fix all breakpoints showing as unresolved in HHVM debugger
Summary: I made a change to fix an issue where unresolved breakpoints were showing as resolved by confirming with the debugger backend before showing the breakpoint as resolved. Unfortunately this breaks a hack that was in place for HHVM where the files are loaded lazily and the Chrome devtools have no idea if the bp is resolvable or not at the time it's set. Reverting to the old behavior of just assuming they are resolved. This is not ideal but it's less broken than always showing as unresolved, which basically breaks HHVM debugging. We likely cannot fix this properly until we remove the Chrome dev tools.

Reviewed By: shushz

Differential Revision: D4777788

fbshipit-source-id: 3a3118e6f10b8a109ff369c5c18aaf2a423eaf3b