Fix flow not found on Windows with npm flow-bin package
Summary: This only repros if Flow is installed with the npm flow-bin package (installing flow.exe somewhere in PATH works fine). The flow-bin package places a flow.cmd script in the npm root directory, which is added to PATH. Nuclide tries to find flow doing `where flow', which will return 2 entries - a "flow" (which is a shell script for linux) and "flow.cmd" which is the wrapper we need to invoke. We take the first one, which is always the wrong one. This change explicitly tries to find flow.cmd if on Windows. If found, use that. Otherwise it will fallback to using whatever "where flow" returns (which will be flow.exe if that's on the path).

Reviewed By: nmote

Differential Revision: D4801522

fbshipit-source-id: 762de637d0a8011e1d0f3e051770ab225a171333