show icons in OutlineView
Summary:
The OutlineView traditionally shows language-keyword declarations like "public static function foo(params)". But that takes up a lot of space. This diff enables outline providers to show an icon in front of their text.

This diff also uses icons for LSP's outline provider. That's handy because LSP doesn't provide the language-keyword declaration: it only provides the SymbolKind enumeration (enough to figure out an icon, the same icon as used in SymbolSearch).

The next step will be to add hover-support to the Outline View. That can be done with vanilla LSP. It will let you hover over an item in the outline-view to see the exact language declaration, plus any docblock.

(Beyond that, we'll augment LSP itself to include parameters in the outline view, and to include public/private visibility, because those things are useful to see directly in the outline-view without needing to hover. But I'm pushing as far as I can with vanilla LSP so we get best possible UX from all other LSP languages, not just Hack.)

Reviewed By: peterhal

Differential Revision: D4819924

fbshipit-source-id: 665f0b125e79939f6cadd16e464b093f09084639