Retry subscription message if diagnostics do not appear
Summary:
This is a dirty hack. I am also investigating the root cause of this issue, but it has escaped me for long enough that I no longer want to block rollout of this feature on it. I have added additional logging in Flow, which will be available in a future version. Hopefully that provides enough information to pinpoint the cause.

Basically, sometimes the subscribe message gets lost somewhere between the Flow client and the Flow server. Since the Flow server is supposed to report the current state of diagnostics immediately upon receiving a subscribe request, it's a safe bet that if we haven't gotten a response after 5 seconds that something went wrong. Plus, the subscribe message is idempotent so sending an extra isn't harmful.

Reviewed By: hansonw

Differential Revision: D4823919

fbshipit-source-id: b9eaf82cdd9ae30c05fb34bd1dbbc708ea96f32c