Add an API for Markdown-based datatips
Summary:
The LSP's `hover` API (equivalent of datatip) works with Markdown strings and code snippets.
https://github.com/Microsoft/language-server-protocol/blob/master/protocol.md#textDocument_hover

This is actually quite convenient for external providers as this alleviates the React requirement.

Adopt this API (`MarkedString`) and use it for typehints.
I actually just moved the type hint UI into datatips in order to easily implement the code snippet part.

Here's a screenshot of the sample: https://pxl.cl/73Wm

Reviewed By: jgebhardt

Differential Revision: D4812360

fbshipit-source-id: d191aeac5f0c92e5c056a3e191c73a4cd2d6f199