respect the server capabilities
Summary:
LSP servers say what methods they support through capabilities.

We should be checking those capabilities, and not attempting to invoke the method if the server doesn't support the feature.

Note: just for completeness, here are the remaining server capabilities which we have no reason to check because Nuclide doesn't yet use them in any way:

  signatureHelpProvider?: SignatureHelpOptions;
  codeActionProvider?: boolean;
  codeLensProvider?: CodeLensOptions;
  documentOnTypeFormattingProvider?: DocumentOnTypeFormattingOptions;
  renameProvider?: boolean;
  documentLinkProvider?: DocumentLinkOptions;
  executeCommandProvider?: ExecuteCommandOptions;

  // This one is needed by Nuclide, but we'll address it in
  // a future diff because it's complicated.
  textDocumentSync?: TextDocumentSyncOptions | number;

Reviewed By: nmote

Differential Revision: D4812399

fbshipit-source-id: 142428523f429d04e5778f86144e022152234c54