Fixing bad messaging in PHP/Hack debugger
Summary:
Fixing some error messaging shown when the debugger gets a bad XML message from the xdebug extention on the PHP server. This can happen if we can't parse the message, or if the message is unexpected (debugger state machine somehow got out of sync with the connection's state in HHVM). When this happens, the debugger panics and disconnects from that particular PHP connection. The current message makes it sound like the entire debug session is being aborted though - when in reality it's just the one connection. The debugger remains attached to any other outstanding PHP requests, and any new ones that come in.

Fixing the messaging because this behavior is misleading and confusing.

Reviewed By: aadisriram, johnislarry

Differential Revision: D4833673

fbshipit-source-id: 8bf6f0cbe3dc13b9c1d04d0d8f576f160eb057af