Use `createPackage()` for file tree activation
Summary: Though it isn't obvious, this also switches the activation logic to occur during the `initialize()` hook (before deserialization). This will fix deserializer errors once we migrate to native workspace views which invokes deserializers when deserializing the workspace (unlike our workspace-views package which does it when the package is activated).

Reviewed By: hansonw

Differential Revision:
D4830187

Tags: nuclide-workspace-views

fbshipit-source-id: 2b8b91e14935b7d526aa6d8f7a90c0c24833278d