Removing diff-view from public release of Nuclide
Summary:
Move diff-view related stuff to internal only version

Changelog:
We are removing diff-view features from the public release of Nuclide, based on feedback and the features offered there has been very low usage outside of Facebook. The lack of support for 'git' and how tied down the diff view is to specific internal requirements and features it leaves us with a product that doesn't meet the quality bar we'd like to set for Nuclide. In order to improve on this, we are moving these features to a Facebook internal version of Nuclide for the foreseeable future. This will enable us to iterate quickly, clean up tech debt and polish the experience we are trying to build for source control without negatively impacting the overall performance in the public version of Nuclide.

Reviewed By: mostafaeweda

Differential Revision: D4836924

fbshipit-source-id: 8f16c236b935ce5f768e858238ccea675349a1cd