PHP Debugger: fix continuing to break on exceptions even after the "pause on exceptions" box is unchecked
Summary: What happens here is while the pause on exceptions box is checked, multiple requests can hit an exception and queue up waiting to break into the debugger. The user then unchecks the box and hits "resume", expecting all requests to resume, but since several are waiting to notify of a break-in already, all those notifications still come in as breaks. This isn't the behavior anyone is expecting.

Reviewed By: shushz

Differential Revision: D4846477

fbshipit-source-id: 38f88aaa3fc698c5447d7f98215fda21543d9e09