Removing source-control-sidebar from public release of Nuclide
Summary:
Source control sidebar is non-useful to the open source release of Nuclide as it doesn't work with `git`.

Changelog: We are removing source control sidebar from the public release of Nuclide as it's almost not useful outside of Facebook as it lacks 'git' support and depends on watchman.

Reviewed By: jgebhardt

Differential Revision:
D4848157

Tags: nuclide-source-control, nuclicator

fbshipit-source-id: 53a602e8a3920835a2b64fc7ddaeded68e5ba66f