HgRepository exposes Observables for head+stack changes too
Summary:
HgRepositoryClient used to expose only a legacy-style API for the list of uncommitted files. ("legacy" in this sense means it wasn't an observable; it was just a cache that gets updated by an RPC call to "hg status" plus an event that fires after that call has finished).

I've updated it to also expose this as an observable.

And I've added two other observables: one that includes changed files in the current head, and one that includes changed files in the entire stack.

Reviewed By: mostafaeweda

Differential Revision: D4739073

fbshipit-source-id: b2405e1136c6f7e3a0dae1731f424fd2bc85b2b0