Add progress indicator to Uncommitted/Head/Stack dropdown
Summary:
For "uncommitted changes", the system already keeps this up to date, and when the user chooses to view uncommitted changes in the File Tree then it's pretty much instantaneous. But for "head changes" and "stack changes", both of those involve an RPC+Mercurial delay of several seconds before the change-list gets updated.

So I've made FileTreeActions communicate while it's in this delay state, and FileTreeStore stores the fact, and FileTreeSidebarComponent renders it with a spinner.

Reviewed By: mostafaeweda

Differential Revision: D4739291

fbshipit-source-id: 3fb53c836dee18642e4e056958d17b94c4bf7b7c