PHP Debugger: Fix triggering async-break while there are no active requests hangs debugger forever.
Summary:
Currently if you trigger an async-break (hit the pause button in the UI) when there are no active PHP requests hangs the debugger. What happens is the front-end issues a pause instruction, which is normally forwarded to all active connections. When a copnnection responds that it has broken in, the response is sent to the front-end, which causes the UI to update the debugger state and request call stacks, scopes, etc.

If there are no requests, there's nobody to ever issue this response, and the UI waits for it forever.  When new requests come in while the debugger is paused, the back-end intentionally doesn't inform the UX because this causes weird behavior where new requests appear spontaneously while the server is supposed to be paused. This is a problem though because the front-end is effectively hung until you kill the debugger.

This change prints a warning to the console letting you know that the debugger is waiting for a request to break. When a request comes in, it will break in. The new request will be broken immediately after it starts, so it won't have a call stack or any scopes (it's not in a frame yet). This is kind of odd since you can't do much with it, but it unhangs the UI and is technically what the user asked for - a new request should not proceed to its first execution frame if the user has already hit "break all" in the UI.

Reviewed By: mostafaeweda

Differential Revision: D4848240

fbshipit-source-id: 09f2874efdf11578924209b1b1d684da4484d8e9