Fix exception when setting breakpoint, and issue where clicking in the gutter sets unwanted breakpoints.
Summary: This fixes an issue where clicking the line number next to a line to select the whole line in Atom actually sets an unwanted breakpoint. The bp should only be set if you click the breakpoint icon.  This also catches an exception that can be thrown if the mouse coordinates don't resolve to a line when setting the bp.

Reviewed By: shushz

Differential Revision: D4853232

fbshipit-source-id: 0bea0c677dc489b18b96b7fd312963eddf67e4ff