Java debugger - show missing source 2/3
Summary:
If the Java debugger cannot locate source for a frame, there are a few features that won't work correctly (especially evaluate on call frame). Currently the Java debugger appends the string "(source file not found)" to the frame name, which is kind of clunky and not as helpful. This stack will pass a flag indicating source is missing all the way through to the front-end, which will display an icon and let the user act on the error by clicking to specify where the source is located.

CL 2: This moves away from using the custom debugger buttons (which was kind of a hack) to adding explicit support for being able to configure source paths. A debugger can specify if it supports this to opt into the new behavior (again, currently only opting Java in, but we can add others if appropriate)

Depends on D4862181

Reviewed By: mostafaeweda

Differential Revision: D4862220

fbshipit-source-id: a64918575d301988aa6222bd84aa0b3338abf568