Make Clang server creation synchronous
Summary:
Now that `RpcProcess` is async again, move the async work (fetching flags and arguments for the server) inside it.

This cleans up a lot of the server management code, as creating a `ClangServer` can be synchronous.

This also allows us to easily have fine-grained state of the server. The server now has several states:

1. finding compilation flags
2. compiling
3. ready
4. disposed

This fixes an annoying bug where language services would block on the first stage (finding compilation flags) due to the asynchronous server creation.

Now, most requests automatically bounce unless the server is in the 'ready' state.

Reviewed By: matthewwithanm

Differential Revision: D4873357

fbshipit-source-id: 091d4de42a2b2b0df18deca7b0c2a42d260af9af