Use a special codepath for performing a fuzzy-filename search in EdenFS.
Summary:
The standard logic for the `nuclide-fuzzy-filename-provider` is not friendly
to EdenFS because, on startup, it enumerates all of the files under the project
root and stores them in memory.

This would defeat the benefits of a virtual filesystem, so we must diverge from
this behavior as early in possible in `FuzzyFileSearchService.js` such that we
never kick off this scan.

This is one of a number of issues with how Nuclide uses Watchman today that will
have to be cleaned up to get the full benefit of Eden.

Reviewed By: hansonw

Differential Revision: D4882798

fbshipit-source-id: 5dd0c029e5950eb81912f1556049bd3714141aeb