Fix sroll stopping in file-tree
Summary:
When scrolling in the file-tree past a section with file names wider when the other parts the scrolling stops abruptly. This does not happen with inertial scrolling, but happens constantly when scrolled by dragging the scrollbar handle. Probably this happens because (due to the windowing nature of the file-tree) the contents of the scrolled element suddenly become wider and this confuses Chrome.

In my tests, setting `overflow` to `scroll` in the panel component scroller fixes this extremely annoying issue, but I don't know whether there are any implications to such change.

Reviewed By: jgebhardt

Differential Revision: D4893930

fbshipit-source-id: 7657c9f16838ed1199924770425de85b35115f9b